# api/openapi.yaml
openapi: 3.1.0
...
info:
  title: Management API
  version: 1.0.0
  summary: The versioned management API — the single door every client (web UI, CLI, MCP server) uses to read and change platform state.
  description: >-
    This document is the source of truth every generated client (TypeScript
    for the web UI, Go for the CLI and MCP server) is produced from. Every
    operation lives under the `/api/v1` prefix. Cross-cutting conventions
    (error shape, optimistic concurrency, pagination, idempotency, trace
    propagation, MCP tag curation, the label-selector grammar, and the
    evolution policy) are specified normatively in `contracts/api-1.md`;
    this document implements those conventions as reusable components and
    applies them to every operation below. `scope-nodes` and `automations`
    are worked fully end to end (schemas, every verb, every error response);
    the remaining operations are SHAPE stubs — path, summary, tag and
    parameters, with the response schema reserved for a later minor.

    Every operation below is served by a real handler, and every route the
    server mounts appears below: the two sets are diffed against each other
    on every build (`internal/app/api`'s surface test), so this document
    cannot promise a generated client an operation that 404s, and a handler
    cannot be added that no client can discover and no schema constrains.
  license:
    name: GNU Affero General Public License v3.0 only
    identifier: AGPL-3.0-only
servers:
...
security:
  - SessionCookie: []
  - ApiKey: []
paths:
...
  /automations/{automation_id}/run:
    parameters:
      - name: automation_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Ulid"
    post:
      operationId: runAutomation
      summary: Run an automation now
      description: >-
        Runs the automation's action sequence NOW, for real, subject to its
        declared mode's normal mode-evaluation rules and to its own
        conditions: a `device_command` reaches the physical device down the
        owning relay's connection, and a signage action writes the targeted
        screens' program override. The response reports what it did, target by
        target. Pass `dry_run: true` to evaluate everything and withhold every
        effect. A mutating POST outside plain resource creation — accepts
        Idempotency-Key so a client's retry-on-timeout cannot double-fire the
        run.
      tags: [automations, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AutomationRunRequest"
      responses:
        "200":
          description: The run completed; the response carries its disposition and every effect it produced.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AutomationRunResult"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "429":
          $ref: "#/components/responses/TooManyRequests"
...
  /webhook-endpoints:
    get:
      operationId: listWebhookEndpoints
      summary: List webhook endpoints
      description: Returns a keyset-paginated, selector-filterable page of registered outbound webhook endpoints.
      tags: [webhook-endpoints, "mcp:read"]
      parameters:
        - $ref: "#/components/parameters/CursorParam"
        - $ref: "#/components/parameters/LimitParam"
        - $ref: "#/components/parameters/SelectorParam"
        - $ref: "#/components/parameters/TraceIdParam"
      responses:
        "200":
          description: A page of webhook endpoints.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpointListResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          $ref: "#/components/responses/TooManyRequests"
    post:
      operationId: createWebhookEndpoint
      summary: Register a webhook endpoint
      description: >-
        Registers an endpoint durable events are delivered to. The endpoint
        receives nothing until a signing secret is installed
        (`rotateWebhookSigningSecret`): an unsigned POST is not a delivery
        `events/1` defines.
      tags: [webhook-endpoints, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointCreate"
      responses:
        "201":
          description: The registered webhook endpoint.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
  /webhook-endpoints/{webhook_endpoint_id}:
    parameters:
      - name: webhook_endpoint_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Ulid"
    get:
      operationId: getWebhookEndpoint
      summary: Read a webhook endpoint
      tags: [webhook-endpoints, "mcp:read"]
      parameters:
        - $ref: "#/components/parameters/TraceIdParam"
      responses:
        "200":
          description: The webhook endpoint.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateWebhookEndpoint
      summary: Update a webhook endpoint
      description: Partial update. Requires If-Match against the endpoint's current ETag/revision.
      tags: [webhook-endpoints, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IfMatchParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookEndpointUpdate"
      responses:
        "200":
          description: The updated webhook endpoint.
          headers:
            ETag:
              $ref: "#/components/headers/ETag"
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookEndpoint"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
    delete:
      operationId: deleteWebhookEndpoint
      summary: Delete a webhook endpoint
      description: >-
        Deletes the registration. The endpoint's delivery progress and its
        sealed signing secret are destroyed with it — a deleted endpoint leaves
        no credential material behind.
      tags: [webhook-endpoints, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IfMatchParam"
        - $ref: "#/components/parameters/TraceIdParam"
      responses:
        "204":
          description: Deleted.
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "412":
          $ref: "#/components/responses/PreconditionFailed"
        "428":
          $ref: "#/components/responses/PreconditionRequired"
  /webhook-endpoints/{webhook_endpoint_id}/signing-secret:
    parameters:
      - name: webhook_endpoint_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Ulid"
    post:
      operationId: rotateWebhookSigningSecret
      summary: Install or rotate a webhook endpoint's signing secret
      description: >-
        Installs the secret every delivery to this endpoint is signed with
        (`events/1` EVT-151). The first call arms the endpoint — until it lands
        the endpoint receives nothing. Every later call is a rotation: the
        secret being replaced stays acceptable for a stated overlap window
        (EVT-158), during which a delivery carries a second signature computed
        under it, so a receiver has time to adopt the new secret before the old
        one stops working. The secret is supplied by the caller and appears in
        no response.
      tags: [webhook-endpoints, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WebhookSigningSecretRequest"
      responses:
        "200":
          description: The rotation's timing. Never the secret.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSigningSecretRotation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/UnprocessableContent"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /webhook-endpoints/{webhook_endpoint_id}/enable:
    parameters:
      - name: webhook_endpoint_id
        in: path
        required: true
        schema:
          $ref: "#/components/schemas/Ulid"
    post:
      operationId: enableWebhookEndpoint
      summary: Re-enable an auto-disabled webhook endpoint
      description: >-
        `events/1` EVT-154 disables an endpoint whose deliveries have failed
        past a bounded consecutive-failure count, and a disabled endpoint
        receives no further attempts until an operator re-enables it. This is
        that act. The endpoint's delivery cursor is left exactly where it was,
        so it resumes from where it stopped rather than skipping what it never
        received (EVT-155). Enabling an already-active endpoint is not an error.
      tags: [webhook-endpoints, "mcp:act"]
      parameters:
        - $ref: "#/components/parameters/IdempotencyKeyParam"
        - $ref: "#/components/parameters/TraceIdParam"
      responses:
        "200":
          description: The endpoint's delivery state after re-enabling.
          headers:
            Trace-Id:
              $ref: "#/components/headers/TraceIdResponse"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookDeliveryState"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    SessionCookie:
      type: apiKey
      in: cookie
      name: session
      description: Browser session cookie issued by /auth/login. Mutating requests additionally require the platform's CSRF token.
    ApiKey:
      type: http
      scheme: bearer
      bearerFormat: "wv_<random>, scoped to a principal and role"
      description: API-key principal (user-issued or internal-service), presented as a bearer token.
  parameters:
...
    IdempotencyKeyParam:
      name: Idempotency-Key
      in: header
      required: false
      description: Client-generated opaque replay key, scoped to (principal, method, path). Optional; strongly recommended on any POST a client might retry.
      schema:
        type: string
        minLength: 1
        maxLength: 255
    TraceIdParam:
      name: Trace-Id
      in: header
      required: false
      description: Caller-supplied trace ID (ULID- or UUID-class, 20-36 chars). A non-conforming value is discarded and replaced server-side; the request still proceeds.
      schema:
        $ref: "#/components/schemas/TraceId"
...
  headers:
...
    TraceIdResponse:
      description: The trace ID that correlates this request across every component it touched — either the caller-supplied value (if it validated) or a freshly generated one.
      required: true
      schema:
        $ref: "#/components/schemas/TraceId"

  # Note: the Deprecation (RFC 9745) and Sunset (RFC 8594) response headers
  # this contract's evolution policy uses (contracts/api-1.md, API-082) are
  # not modeled as reusable components here — nothing in this document is
  # deprecated yet, so there is no operation to attach them to without
  # faking a deprecation. contracts/api-1.md is their normative source; a
  # future minor wires them onto whichever operation first deprecates.
  responses:
    BadRequest:
      description: The request was malformed (validation failure, invalid selector, invalid cursor).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Unauthorized:
      description: No valid principal was presented.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Ulid:
      type: string
      description: A 26-character Crockford-base32 identifier, lexicographically sortable by creation time.
      pattern: "^[0-9A-HJKMNP-TV-Z]{26}$"
      examples: ["01J8Z3K4N5P6Q7R8S9T0V1W2X3"]
...
    AutomationRunRequest:
      type: object
      description: Optional trigger-context override for this manual run.
      additionalProperties: false
      properties:
        context:
          type: object
          additionalProperties: true
        dry_run:
          type: boolean
          default: false
          description: >-
            When `true`, the run is evaluated to completion — conditions,
            branch selection, target resolution — but every effect is WITHHELD:
            no device command is dispatched and no screen override is written.
            The response reports exactly what a real run would have done. The
            default is `false`: run-now ACTS.
...